REFACTOR: Move Eggnog things to own submodule.
