+++ q2_types_genomics/eggnog/__init__.py
@@ -7,15 +7,31 @@
 # ----------------------------------------------------------------------------
 
 
+import importlib
+
 from ._format import (
         EggnogRefBinFileFmt, EggnogRefTextFileFmt, EggnogRefDirFmt,
-        EggnogOutputDirFmt, PfamDirFmt, DiamondRefDirFmt,
+        EggnogOutputDirFmt, PfamDirFmt, DiamondRefDirFmt, MMseqsDirFmt,
+        ArbitraryHeaderTSVFmt, ArbitraryHeaderTSVDirFmt
         )
 
-from ._types import (Eggnog, ReferenceDB)
+from ._type import (
+        Orthologs, Seed, Eggnog, ReferenceDB, Annotation, NOG, KEGG, OG,
+        EggnogDB,
+        )
 
 __all__ = [
         'EggnogRefBinFileFmt', 'EggnogRefTextFileFmt', 'EggnogRefDirFmt',
         'EggnogOutputDirFmt', 'Eggnog', 'ReferenceDB', 'PfamDirFmt',
-        'DiamondRefDirFmt'
+        'DiamondRefDirFmt',  'MMseqsDirFmt', 'Orthologs', 'Seed', 'ArbitraryHeaderTSVFmt',
+        'ArbitraryHeaderTSVDirFmt', 'EggnogDB', 'KEGG', 'OG', 'DiamondDB', 'MMseq2DB',
+        'NOG',
         ]
+
+importlib.import_module('q2_types_genomics.eggnog._transformer')
+importlib.import_module('q2_types_genomics.eggnog._validator')
+    #'ArbitraryHeaderTSVDirFmt',
+     #'BinaryReferenceDBFmt',
+    #'BinaryReferenceDBDirFmt',
+    #ArbitraryHeaderTSVDirFmt, #BinaryReferenceDBFmt, BinaryReferenceDBDirFmt,
+
