+++ q2_types_genomics/eggnog/_methods.py
@@ -5,3 +5,7 @@
 #
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
+
+from ..eggnog import OrthologDirFormat, Ortholog, Seed
+
+
