+++ q2_types_genomics/feature_data/_type.py
@@ -6,14 +6,10 @@
 # The full license is in the file LICENSE, distributed with this software.
 # ----------------------------------------------------------------------------
 
+from qiime2.core.type import SemanticType
 from q2_types.feature_data import FeatureData
 
-from q2_types_genomics.eggnog import EggnogRefDirFmt
-from q2_types_genomics.feature_data._format import (
-    MAGSequencesDirFmt, ArbitraryHeaderTSVDirFmt,
-    BinaryReferenceDBDirFmt,
-    )
-from qiime2.core.type import SemanticType
+from q2_types_genomics.feature_data._format import MAGSequencesDirFmt
 from ..plugin_setup import plugin
 
 
@@ -24,50 +20,3 @@ plugin.register_semantic_type_to_format(
     FeatureData[MAG],
     artifact_format=MAGSequencesDirFmt
 )
-
-# NOG stuff
-NOG = SemanticType('NOG', variant_of=FeatureData.field['type'])
-plugin.register_semantic_types(NOG)
-plugin.register_semantic_type_to_format(
-    semantic_type=FeatureData[NOG],
-    artifact_format=ArbitraryHeaderTSVDirFmt
-)
-
-# KEGG stuff
-KEGG = SemanticType('KEGG', variant_of=FeatureData.field['type'])
-plugin.register_semantic_types(KEGG)
-plugin.register_semantic_type_to_format(
-        semantic_type=FeatureData[KEGG],
-        artifact_format=ArbitraryHeaderTSVDirFmt
-)
-
-# OG stuff
-OG = SemanticType('OG', variant_of=FeatureData.field['type'])
-plugin.register_semantic_types(OG)
-plugin.register_semantic_type_to_format(
-        semantic_type=FeatureData[OG],
-        artifact_format=ArbitraryHeaderTSVDirFmt
-)
-
-# types for downloading the databases eggnogmapper is using.
-DiamondDB = SemanticType('DiamondDB', variant_of=FeatureData.field['type'])
-plugin.register_semantic_types(DiamondDB)
-plugin.register_semantic_type_to_format(
-        FeatureData[DiamondDB],
-        artifact_format=BinaryReferenceDBDirFmt
-)
-
-MMseq2DB = SemanticType('MMseq2DB', variant_of=FeatureData.field['type'])
-plugin.register_semantic_types(MMseq2DB)
-plugin.register_semantic_type_to_format(
-        FeatureData[MMseq2DB],
-        artifact_format=BinaryReferenceDBDirFmt
-)
-
-EggnogDB = SemanticType('EggnogDB')
-plugin.register_semantic_types(EggnogDB)
-plugin.register_semantic_type_to_format(EggnogDB, EggnogRefDirFmt)
-# plugin.register_semantic_type_to_format(
-#         EggnogDB,
-#         artifact_format=BinaryReferenceDBDirFmt
-# )
